// ui/src/api/generated/model/checklistAnswerEntry.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * Whirlpool Inspection Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.0.3
 */

export interface ChecklistAnswerEntry {
  /** @minimum 1 */
  id: number;
  value: string;
  image_path?: string[];
  remarks?: string | null;
}