// ui/src/api/generated/model/inspectionMetadataResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * Whirlpool Inspection Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.0.3
 */
import type { InspectionDropdownOption } from './inspectionDropdownOption';

export interface InspectionMetadataResponse {
  inspection_types: InspectionDropdownOption[];
  warehouses: InspectionDropdownOption[];
  plants: InspectionDropdownOption[];
  damage_types: InspectionDropdownOption[];
  damage_severities: InspectionDropdownOption[];
  damage_causes: InspectionDropdownOption[];
  damage_grades: InspectionDropdownOption[];
  review_statuses: InspectionDropdownOption[];
}